let users.create and users.update take arbitrary argument and set them on the user
