+++ test/testsuite/tests/db.js
@@ -782,215 +782,3 @@ exports['newUUID - cache concurrency'] = function (test)
         test.done();
     });
 };
-
-exports['listUsers'] = function (test)
-{
-    async.waterfall([
-        function(callback) {
-            users.list(function(err, users) {
-                test.equal(err, undefined, 'listUsers has no error');
-                test.notEqual(users, undefined, 'users are defined');
-                callback();
-            });            
-        }
-    ], function () {
-        test.done();
-    });
-    
-};
-
-
-
-
-
-
-
-var deleteUser = function(username, test, callback) {
-    users.delete(username, function(err) {
-        if (err) {
-            test.done(err);
-        } else {
-            callback();
-        }
-    });    
-};
-
-exports['createUser'] = function (test)
-{
-    test.expect(5);
-
-    async.waterfall([
-        function (callback) {
-            users.create('wizard_of_oz', 'password', function(err, user) {
-                if (err) { return test.done(err); }
-                
-                test.equal(err, undefined, 'signup has no error');
-                test.notEqual(user, undefined, 'signup returns a value');
-                test.equal(user.ok, true, 'signup returns okay');
-                callback();
-            });
-        },
-        function (callback) {
-            users.create('wicked_with_of_the_east', 'password', ['witch'], function(err, user) {
-                if (err) { return test.done(err); }
-                
-                users.get('wicked_with_of_the_east', function(err, user, options) {
-                    if (err) { return test.done(err); }
-                    
-                    test.equal(user.roles[0], "witch");
-                    callback();                    
-                });
-            });            
-        },
-        function (callback) {
-            users.create('tin_woodman', 'password', ['_admin'], function(err, user) {
-                if (err) { return test.done(err); }
-                
-                db.request({
-                    type: 'GET',
-                    url: '/_config/admins/tin_woodman',
-                    contentType: 'application/json'                
-                }, function(err, admin) {
-                    if (err) { return test.done(err); }
-                    
-                    test.notEqual(admin, undefined, 'tin_woodman should be an admin');
-                    callback();
-                });
-            });
-        },
-        function (callback) {
-            deleteUser('wizard_of_oz', test, function() {
-                deleteUser('tin_woodman', test, function() {
-                    deleteUser('wicked_with_of_the_east', test, callback);
-                });
-            });
-        }
-    ], function () {
-        test.done();
-    });
-};
-
-
-exports['updateUser'] = function (test)
-{
-    test.expect(5);
-
-    async.waterfall([
-        function (callback) {
-            // if new role is admin user should be admin after (1)
-            users.create('wizard_of_oz', 'password', ['wizard'], function(err, user) {
-                if (err) { return test.done(err); }
-
-                users.update('wizard_of_oz', 'password', ['_admin'], function(err, user) {
-                    if (err) { return test.done(err); }
-
-                    db.request({
-                        type: 'GET',
-                        url: '/_config/admins/wizard_of_oz',
-                        contentType: 'application/json'                
-                    }, function(err, admin) {
-                        if (err) { return test.done(err); }
-                    
-                        test.notEqual(admin, undefined, 'wizard_of_oz should be an admin');
-                        callback();
-                    });
-                });
-            });            
-        },
-        function (callback) {
-            // if user was admin but new role isn't, user should not be admin after (1)
-            users.create('wicked_with_of_the_east', 'password', ['_admin'], function(err, user) {
-                if (err) { return test.done(err); }
-        
-                users.update('wicked_with_of_the_east', 'password', ['wizard'], function(err, user) {
-                    if (err) { return test.done(err); }
-        
-                    db.request({
-                        type: 'GET',
-                        url: '/_config/admins/wicked_with_of_the_east',
-                        contentType: 'application/json'
-                    }, function(err, admin) {
-                        test.equal(admin, undefined, 'wicked_with_of_the_east should not be an admin');
-                        callback();
-                    });
-                });
-            });
-        },
-        function (callback) {
-            // user should have new roles afterwards (1)
-            users.create('tin_woodman', 'password', ['companion'], function(err, user) {
-                if (err) { return test.done(err); }
-        
-                users.update('tin_woodman', 'password', ['wizard'], function(err, user) {
-                    if (err) { return test.done(err); }
-        
-                    users.get('tin_woodman', function(err, user, options) {
-                        if (err) { return test.done(err); }
-        
-                        test.equal(user.roles[0], "wizard");
-                        callback();                    
-                    });
-                });
-            });            
-        },
-        function (callback) {
-            // user should have new password afterwards (1)
-            users.create('scarecrow', 'password', ['companion'], function(err, user) {
-                if (err) { return test.done(err); }
-        
-                users.get('scarecrow', function(err, user) {
-                    if (err) { return test.done(err); }
-
-                    var old_password = user.password_sha;
-                    
-                    users.update('scarecrow', 'newpassword', ['companion'], function(err, user) {
-                        if (err) { return test.done(err); }
-
-                        users.get('scarecrow', function(err, user) {
-                            if (err) { return test.done(err); }
-
-                            test.notEqual(user.password_sha, old_password);
-                            callback();
-                        });
-                    });                    
-                });
-            });
-        },
-        function (callback) {
-            // empty password should not change password (1)
-            users.create('cowardly_lion', 'password', ['companion'], function(err, user) {
-                if (err) { return test.done(err); }
-        
-                users.get('cowardly_lion', function(err, user) {
-                    if (err) { return test.done(err); }
-                    
-                    var old_password = user.password_sha;
-                    
-                    users.update('cowardly_lion', '', ['companion'], function(err, user) {
-                        if (err) { return test.done(err); }
-
-                        users.get('cowardly_lion', function(err, user) {
-                            if (err) { return test.done(err); }
-
-                            test.equal(user.password_sha, old_password);
-                            callback();
-                        });
-                    });
-                });                
-            });            
-        },
-        function (callback) {
-            deleteUser('wicked_with_of_the_east', test, function() {
-                deleteUser('tin_woodman', test, function() {
-                    deleteUser('scarecrow', test, function() {
-                        deleteUser('cowardly_lion', test, function() {                    
-                            deleteUser('wizard_of_oz', test, callback);
-                        });
-                    });
-                });
-            });
-        }
-    ], function () {
-        test.done();
-    });
-};
